Refactor primary relay selector to separate component
