+++ packages/react-libp2p-debug/src/components/SelfInfo.jsx
@@ -1,19 +1,15 @@
 import React, { useCallback, useEffect, useState } from 'react';
 
 import { getPeerSelfInfo, getPseudonymForPeerId } from '@cerc-io/libp2p-util';
-import { Box, Button, FormControl, Grid, InputLabel, MenuItem, Paper, Select, Table, TableBody, TableCell, TableContainer, TableRow, Typography } from '@mui/material';
+import { Box, Grid, Paper, Table, TableBody, TableCell, TableContainer, TableRow, Typography } from '@mui/material';
 
 import { DEFAULT_REFRESH_INTERVAL, THROTTLE_WAIT_TIME } from '../constants';
 import { useThrottledCallback } from '../hooks/throttledCallback';
-import { CustomRelayDialog } from './CustomRelayDialog';
+import { PrimaryRelaySelector } from './PrimaryRelaySelector';
 
 const STYLES = {
   selfInfoHead: {
     marginBottom: 1/2
-  },
-  primaryRelaySelect: {
-    marginRight: 1,
-    minWidth: 200
   },
   nodeStartedTableCell: {
     width: 150
@@ -29,44 +25,11 @@ export function SelfInfo ({
   ...props
 }) {
   const [selfInfo, setSelfInfo] = useState();
-  const [openCustomRelayDialog, setOpenCustomRelayDialog] = useState(false);
-  const [primaryRelay, setPrimaryRelay] = useState(localStorage.getItem('primaryRelay') ?? '')
-  const [customRelay, setCustomRelay] = useState(localStorage.getItem('customRelay'))
 
   const updateInfo = useCallback(() => {
     setSelfInfo(getPeerSelfInfo(node))
   }, [node])
   const throttledUpdateInfo = useThrottledCallback(updateInfo, THROTTLE_WAIT_TIME);
-
-  const handlePrimaryRelaySelectChange = useCallback(event => {
-    // Check if value is null when custom relay option is selected
-    if (event.target.value !== null) {
-      setPrimaryRelay(event.target.value)
-    }
-  }, [])
-
-  const handlePrimaryRelayUpdate = useCallback(() => {
-    // Set selected primary relay in localStorage and refresh app
-    localStorage.setItem('primaryRelay', primaryRelay);
-    window.location.reload(false);
-  }, [primaryRelay])
-
-  const handleCustomOptionClick = useCallback(event => {
-    setOpenCustomRelayDialog(true);
-  }, [])
-
-  const handleCustomRelaySet = useCallback((newCustomRelay) => {
-    setCustomRelay(newCustomRelay);
-    // Set custom relay in localStorage
-    localStorage.setItem('customRelay', newCustomRelay);
-    setPrimaryRelay(newCustomRelay);
-    setOpenCustomRelayDialog(false);
-  }, [])
-
-  const handleCustomRelayCancel = useCallback(() => {
-    setPrimaryRelay(primaryRelay);
-    setOpenCustomRelayDialog(false);
-  }, [primaryRelay])
 
   useEffect(() => {
     if (!node) {
@@ -92,53 +55,10 @@ export function SelfInfo ({
         <Grid item xs />
         {enablePrimaryRelaySupport && (
           <Grid item xs="auto">
-            <Box display="flex" alignItems="flex-end">
-              <FormControl variant="standard" sx={STYLES.primaryRelaySelect} size="small">
-                <InputLabel shrink id="primary-relay-label">Primary Relay</InputLabel>
-                <Select
-                  displayEmpty
-                  labelId="primary-relay-label"
-                  id="primary-label"
-                  value={primaryRelay}
-                  label="Primary Relay"
-                  onChange={handlePrimaryRelaySelectChange}
-                >
-                  <MenuItem value="">{"<random>"}</MenuItem>
-                  {relayNodes.map(relayNode => (
-                    <MenuItem
-                      value={relayNode}
-                      key={relayNode}
-                    >
-                      {relayNode.split('/')[2]}
-                    </MenuItem>
-                  ))}
-                  <MenuItem
-                    value={customRelay}
-                    onClick={handleCustomOptionClick}
-                  >
-                    {"<custom>"}
-                  </MenuItem>
-                </Select>
-              </FormControl>
-              <Button
-                variant="contained"
-                size="small"
-                disabled={primaryRelay === (localStorage.getItem('primaryRelay') ?? '')}
-                onClick={handlePrimaryRelayUpdate}
-                sx={STYLES.primaryRelayButton}
-              >
-                UPDATE
-              </Button>
-            </Box>
+            <PrimaryRelaySelector relayNodes={relayNodes} />
           </Grid>
         )}
       </Grid>
-      <CustomRelayDialog
-        open={openCustomRelayDialog}
-        oldCustomRelay={customRelay}
-        onSet={handleCustomRelaySet}
-        onCancel={handleCustomRelayCancel}
-      />
       <TableContainer component={Paper}>
         <Table size="small">
           <TableBody>
